% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GINAX_terminal}
\alias{GINAX_terminal}
\title{GINAX function}
\usage{
GINAX_terminal(
  Y,
  kinship,
  Z,
  SNPs,
  family,
  offset = NULL,
  FDR.threshold,
  maxiterations,
  runs_til_stop
)
}
\description{
GINAX function
}
\keyword{internal}
