% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinship}
\alias{kinship}
\title{Example Dataset1: kinship}
\format{
A matrix for kinship
}
\source{
Generated for package example
}
\usage{
data(kinship)
}
\description{
This dataset contains kinship matrix.
}
\keyword{datasets}
