% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log.marginal.likelihood}
\alias{log.marginal.likelihood}
\title{likelihood function}
\usage{
\method{log}{marginal.likelihood}(k, x.tilde_m, y.tilde, D_inv, ydinvy, dinvy, g)
}
\description{
likelihood function
}
\keyword{internal}
