% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_marginal_likelihood_null}
\alias{log_marginal_likelihood_null}
\title{likelihood function for null}
\usage{
log_marginal_likelihood_null(y.tilde, D_inv)
}
\description{
likelihood function for null
}
\keyword{internal}
