% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNPs}
\alias{SNPs}
\title{Example Dataset4: SNPs}
\format{
A data frame with 328 rows and 9000 variables
}
\source{
Generated for package example
}
\usage{
data(SNPs)
}
\description{
This dataset contains all SNPs.
}
\keyword{datasets}
