% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_binary}
\alias{Y_binary}
\title{Example Dataset3: Y_binary}
\format{
A vector for binary data
}
\source{
Generated for package example
}
\usage{
data(Y_binary)
}
\description{
This dataset contains response variable (binary data).
}
\keyword{datasets}
