% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_poisson}
\alias{Y_poisson}
\title{Example Dataset2: Y_poisson}
\format{
A vector for poisson data
}
\source{
Generated for package example
}
\usage{
data(Y_poisson)
}
\description{
This dataset contains response variable (count data).
}
\keyword{datasets}
